+++ fr.opensagres.mongodb.ide.ui/src/fr/opensagres/mongodb/ide/ui/viewers/editor/IProgressBarValueProvider.java
@@ -2,5 +2,5 @@ package fr.opensagres.mongodb.ide.ui.viewers.editor;
 
 public interface IProgressBarValueProvider {
 
-	double getProgressBarValue(Object element);
+	int getProgressBarValue(Object element);
 }
